Backend/requirements.txt
fastapi
uvicorn[standard]
httpx
passlib
pydantic
pydantic[email]
PyJWT
python-dotenv
SQLAlchemy